requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.34.0
google-genai==1.63.0
pydantic==2.10.0
python-dotenv==1.0.1